+++ changelog.md
@@ -3,6 +3,7 @@ v5.0.1-beta
 * Updated to Forge 52.0.13
 * Added affinity data for new items and creatures
 * Added sanguine cores for new creature types
+* Added rune enchantment definitions for new mace enchantments
 * Runecarving Table recipes now properly grant expertise
 * Fixed a crash that occurs when a thrown magickal metal trident strikes a creature
 * Fixed a bug preventing the Silk Touch property for Break spells working in certain circumstances